+++ js/length-line.js
@@ -1,0 +1,10 @@
+//Функция сравнения длины строки
+function getLengthLine(string,length){
+  if(string<=length){
+    return 'true';
+  }else{
+    return 'false';
+  }
+}
+
+export {getLengthLine};
