+++ js/random-array-element.js
@@ -1,0 +1,6 @@
+import {getRandomInt} from './random-integer.js';
+
+//Функция по поиску случайного элемента в переданном массиве
+const getRandomArrayElement = (elements) => elements[getRandomInt(0, elements.length - 1)];
+
+export {getRandomArrayElement};
